* Converted itemdb tables to use smallint. Credits to ALZ.
modified   Changelog-Trunk.txt
modified   sql-files/item_db.sql
modified   sql-files/item_db2.sql
added      sql-files/upgrade_svn9135.sql


git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@9135 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ sql-files/item_db2.sql
@@ -22,7 +22,7 @@ CREATE TABLE `item_db2` (
   `weapon_level` tinyint(2) unsigned default NULL,
   `equip_level` tinyint(3) unsigned default NULL,
   `refineable` tinyint(1) unsigned default NULL,
-  `view` tinyint(3) unsigned default NULL,
+  `view` smallint(3) unsigned default NULL,
   `script` text,
   `equip_script` text,
   `unequip_script` text,
